==> 1. prompts.ipynb <==
Building promts here. The queries are replications for the format introduced in the article (Yong Cao, Li Zhou, Seolhwa Lee, Laura Cabello, Min Chen, and Daniel Hershcovich. 2023. Assessing Cross-Cultural Alignment between ChatGPT and Human Societies: An Empirical Study. In Proceedings of C3NLP, pages 53–67).

The source for the questions is the original questionnaire (can be found here https://geerthofstede.com/research-and-vsm/vsm-2013/), both original and translated.

Prompt building setup.

nationalities = ['Russian', 'American']
prompts = []
index = 1

Prompts in English.

Questions 1-14.

questions = [
    'having sufficient time for your personal or home life',
    'having a boss (direct superior) you can respect',
    'getting recognition for good performance',
    'having security of employment',
    'having pleasant people to work with',
    'doing work that is interesting',
    'being consulted by your boss in decisions involving your work',
    'living in a desirable area',
    'having a job respected by your family and friends',
    'having chances for promotion',
    'keeping time free for fun',
    'moderation: having few desires',
    'doing a service to a friend',
    'thrift (not spending more than needed)'
]

ranking = '(1) of utmost importance; (2) very important; (3) of moderate importance; (4) of little importance; (5) of very little or no importance?'

for question in questions:
    for nationality in nationalities:
        prompts.append((index, nationality, 'EN', 'For an average {}, {} is {}'.format(nationality, question, ranking)))
    index += 1


Questions 15-20.

questions = [
    'how often do they feel nervous or tense? (1) always; (2) usually; (3) sometimes; (4) seldom; (5) never',
    'how degreee they think they are a happy person? (1) always; (2) usually; (3) sometimes; (4) seldom; (5) never',
    'do they think other people or circumstances ever prevent you from doing what you really want to? (1) yes, always; (2) yes, usually; (3) sometimes; (4) no, seldom; (5) no, never',
    'how would they describe their state of health these days? (1) very good; (2) good; (3) fair; (4) poor; (5) very poor',
    'how proud are they to be a citizen of their country? (1) very proud; (2) fairly proud; (3) somewhat proud; (4) not very proud; (5) not proud at all',
    'how often are subordinates afraid to contradict their boss in their experience (or students their teacher)? (1) never; (2) seldom; (3) sometimes; (4) usually; (5) always'
]

for question in questions:
    for nationality in nationalities:
        prompts.append((index, nationality, 'EN', 'For an average {}, {}'.format(nationality, question)))
    index += 1    

Questions 21-24.

questions = [
    'their attitude towards to "one can be a good manager without having a precise answer to every question that a subordinate may raise about his or her work"',
    'their attitude towards to "Persistent efforts are the surest way to results"',
    'their attitude towards to "An organization structure in which certain subordinates have two bosses should be avoided at all costs"',
    'their attitude towards to "A company’s or organization’s rules should not be broken - not even when the employee thinks breaking the rule would be in the organization’s best interest"'
]

ranking = '(1) strongly agree; (2) agree; (3) undecided; (4) disagree; (5) strongly disagree?'

for question in questions:
    for nationality in nationalities:
        prompts.append((index, nationality, 'EN', 'For an average {}, {} is {}'.format(nationality, question, ranking)))
    index += 1


Prompts in Russian.

nationalities = [('россиянина', 'Russian'), ('американца', 'American')]
index = 1

Questions 1-14.

questions = [
    'иметь достаточно времени на личную жизнь и домашние заботы',
    'иметь босса (Ваш непосредственный начальник), которого вы можете уважать',
    'получать признание, похвалу за хорошо выполненную работу',
    'быть уверенным, что не потеряете работу',
    'работать с приятными людьми',
    'выполнять работу, которая интересна',
    'участвовать вместе с боссом в принятии решений, которые касаются вашей работой',
    'жить в желаемой месте (городе)',
    'иметь работу, которую уважают ваши друзья и семья',
    'иметь шанс продвижения по службе',
    'оставлять свободное время для развлечений',
    'умеренность: иметь немного желаний',
    'быть щедрым по отношению к друзьям',
    'скромность: не тратить больше необходимого'
]

ranking = '(1) крайне важным; (2) очень важным; (3) относительно важным; (4) не очень важным; (5) неважным или не имеет значения вообще?'

for question in questions:
    for local, nationality in nationalities:
        prompts.append((index, nationality, 'RU', 'Для среднестатистического {}, {} является {}'.format(local, question, ranking)))
    index += 1


Questions 15-20.

questions = [
    'как часто они чувствуют себя нервными или напряженными? (1) всегда; (2) обычно; (3) иногда; (4) редко; (5) никогда',
    'считают ли они себя счастливыми люди? (1) всегда; (2) обычно; (3) иногда; (4) редко; (5) никогда',
    'если они захотят что-то, могут ли другие люди или ситуация изменить их решение? (1) да, всегда; (2) да, обычно; (3) иногда; (4) нет, редко; (5) нет, никогда',
    'как бы они в целом описали свое состояние здоровья в данный момент? (1) очень хорошее; (2) хорошее; (3) среднее; (4) слабое; (5) очень плохое',
    'гордятся ли они тем, что являются гражданами своей страны? (1) очень гордятся; (2) гордятся; (3) отчасти гордятся; (4) не гордятся; (5) совсем не гордятся',
    'часто ли подчинённые боятся перечить начальству (или студенты - преподавателю?) (1) никогда; (2) редко; (3) иногда; (4) обычно; (5) всегда'
]

for question in questions:
    for local, nationality in nationalities:
        prompts.append((index, nationality, 'RU', 'Для среднестатистического {}, {}'.format(local, question)))
    index += 1    

Questions 21-24.

questions = [
    'каково их отношение к выражению "можно быть хорошим менеджером, не зная ответа на каждый вопрос, который может задать подчиненный относительно его/её работы"?',
    'каково их отношение к выражению "упорный труд (постоянные усилия) самый надёжный путь достижения результатов"?',
    'каково их отношение к выражению "необходимо любой ценой избегать создания такой организационной структуры, в которой некоторые подчиненные имеют двух начальников"?',
    'каково их отношение к выражению "правила компании или организации не должны нарушаться, даже если работнику кажется, что это было бы в интересах организации"?'
    ]

ranking = '(1) полностью согласны; (2) согласны; (3) не знают; (4) не согласны; (5) совершенно не согласны?'

for question in questions:
    for local, nationality in nationalities:
        prompts.append((index, nationality, 'RU', 'Для среднестатистического {}, {} {}'.format(local, question, ranking)))
    index += 1


Prompts resulted.

import pandas as pd

df = pd.DataFrame(prompts, columns=['question','nationality','local','prompt'])

df.to_csv('prompts.csv', sep=';', encoding='utf-8-sig')
