==> 2.1. probing chatgpt.ipynb <==
Probing OpenAI GPT-3.5-turbo-0125 and 4o-2024-05-13 models. Each question to be probed 10 times in a row in isolation from any context.

Docs:
https://platform.openai.com/docs/api-reference/making-requests

import json

with open('api_keys.json') as json_file:
    apikeys = json.load(json_file)

apikey = apikeys['openai']

Calling API, prompbing exactly one prompt each call.

from openai import OpenAI

def get_response_from_chatgpt(prompt, gptmodel):
    client = OpenAI(api_key=apikey)
    response = client.chat.completions.create(
    messages=[
        {
            "role": "user",
            "content": prompt,
        }
    ],
    model=gptmodel)
    return response.choices[0].message.content

import pandas as pd

df_prompts = pd.read_csv('prompts.csv', sep=';', encoding='utf-8-sig').set_index('id')

responses = []
models = ['gpt-4o', 'gpt-3.5-turbo']

for model in models:
    for prompt in df_prompts['prompt']:
        for i in range(10):
            message = get_response_from_chatgpt(prompt, model)
            responses.append((model, prompt, message))

df_responses = pd.DataFrame(responses, columns=['model','question','response'])

df_responses.to_csv('responses_openAI.csv', sep=';', encoding='utf-8-sig')

df_responses
