==> 5.2. locale deviations.ipynb <==
Identifying deviations in responses for the same nationality using different locales.

import pandas as pd

df_responses_gpt = pd.read_csv('responses_parsed_gpt.csv', sep=';', encoding='utf-8-sig')

Descriptive stats for getting min/max for each model.

desc = df_responses_gpt[['question','nationality','local','score_gpt-4o','score_gpt-3.5-turbo','score_yandexgpt']].groupby(['nationality','question','local']).describe()

desc

models = ['score_gpt-4o', 'score_gpt-3.5-turbo', 'score_yandexgpt']

for model in models:
    desc[(model, 'range')] = desc[(model, 'max')] - desc[(model, 'min')]

    desc[[(model, 'mean'),(model, 'std'),(model, 'min'),(model, 'max'),(model, 'range')]][desc[(model, 'range')] >= 2.0].apply(lambda x: round(x,2)).to_csv(f'responses_range_{model}.csv', sep=';', encoding='utf-8-sig')

Defining as outliers cases where difference in scores 2.0 or more between max and min for cross-compared locales, i.e. max(EN) - min(RU).

questions = range(1,25)
nationalities = ['American', 'Russian']

outlying_questions = []

for model in models:
    for question in questions:
        for nationality in nationalities:
            sliced = desc[(model)][['min','max']].loc[(nationality,question)]

            range1 = sliced.loc['EN']['max'] - sliced.loc['RU']['min']
            range2 = sliced.loc['RU']['max'] - sliced.loc['EN']['min']
            
            if range1 >= 2.0 or range2 >= 2.0:
                outlying_questions.append((model.replace('score_',''),question,nationality,max(range1,range2)))

outliers = pd.DataFrame(outlying_questions, columns=['model','question','nationality','range'])

Grouping by to get the number of occurances and list of respective questions with outliers.

outliers.groupby(['model', 'nationality']).agg(
    count=('range', 'count'),
    questions=('question', list)
).reset_index()
