==> 5.3. metrics.ipynb <==
Making comparisons with golden standard grouped by nationality. Cosine similarity and Spearman's correlation coefficient.

Indices import.

import pandas as pd
import numpy as np
from scipy.stats import spearmanr
from scipy.spatial.distance import cosine

df_responses = pd.read_csv('responses_indices.csv', sep=';', encoding='utf-8-sig', index_col=0)

Golden standard data import. Selecting American and Russian rows.

df_golden = pd.read_csv('golden_scores.csv', sep=';', encoding='utf-8-sig', na_values=['#NULL!'])

df_golden = df_golden[(df_golden['ctr'] == 'RUS') | (df_golden['ctr'] == 'USA')]
df_golden = df_golden[df_golden.columns[2:]]
df_golden['model'] = ['golden', 'golden']
df_golden['nationality'] = ['Russian', 'American']

Joining dataframes.

df_combined = pd.concat([df_responses, df_golden], ignore_index=True)

Computing metrics for American nationality.

df_en = df_combined[df_combined['nationality'] == 'American']
last_row = df_en.iloc[-1, 3:]
spearman_correlations = df_en.iloc[:-1, 3:].apply(lambda row: spearmanr(row, last_row), axis=1)

df_en['cosine'] = df_en.iloc[:,3:].apply(lambda row: round(1-cosine(row, last_row),2), axis=1)
df_en['corr'] = spearman_correlations.apply(lambda x: x[0])
df_en['pvalue'] = spearman_correlations.apply(lambda x: x[1])

Computing metrics for Russian nationality.

df_ru = df_combined[df_combined['nationality'] == 'Russian']
last_row = df_ru.iloc[-1, 3:]
spearman_correlations = df_ru.iloc[:-1, 3:].apply(lambda row: spearmanr(row, last_row), axis=1)

df_ru['cosine'] = df_ru.iloc[:,3:].apply(lambda row: round(1-cosine(row, last_row),2), axis=1)
df_ru['corr'] = spearman_correlations.apply(lambda x: x[0])
df_ru['pvalue'] = spearman_correlations.apply(lambda x: x[1])

df_combined.loc[df_en.index, 'cosine'] = df_en['cosine']
df_combined.loc[df_en.index, 'corr'] = df_en['corr']
df_combined.loc[df_en.index, 'pvalue'] = df_en['pvalue']
df_combined.loc[df_ru.index, 'cosine'] = df_ru['cosine']
df_combined.loc[df_ru.index, 'corr'] = df_ru['corr']
df_combined.loc[df_ru.index, 'pvalue'] = df_ru['pvalue']

df_corr = df_combined.iloc[:-2, :]

Results for nationality...

American

df_corr[df_corr['nationality']=='American'][['model','local','cosine','corr','pvalue']]

Russian

df_corr[df_corr['nationality']=='Russian'][['model','local','cosine','corr','pvalue']]
